+++ src/main/java/com/washingtonpost/arc/ans/v0_3/model/Collection.java
@@ -25,11 +25,8 @@ public class Collection extends Content {
     @JsonProperty("short_url")
     private String shortUrl;
 
-    @JsonProperty("title")
-    private String title;
-
-    @JsonProperty("headline")
-    private String headline;
+    @JsonProperty("headlines")
+    private List<Headline> headlines;
 
     @JsonProperty("description")
     private String description;
@@ -87,31 +84,17 @@ public class Collection extends Content {
     }
 
     /**
-     * @return The title of the collection.
+     * @return The headlines for this collection.
      */
-    public String getTitle() {
-        return title;
+    public List<Headline> getHeadlines() {
+        return headlines;
     }
 
     /**
-     * @param title The title of the collection.
+     * @param headlines The headlines for this collection.
      */
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    /**
-     * @return The headline for the collection.
-     */
-    public String getHeadline() {
-        return headline;
-    }
-
-    /**
-     * @param headline The headline for the collection.
-     */
-    public void setHeadline(String headline) {
-        this.headline = headline;
+    public void setHeadline(List<Headline> headlines) {
+        this.headlines = headlines;
     }
 
     /**
@@ -255,7 +238,7 @@ public class Collection extends Content {
                 .append(this.description)
                 .append(this.displayDate)
                 .append(this.editorNote)
-                .append(this.headline)
+                .append(this.headlines)
                 .append(this.contentElements)
                 .append(this.promoImages)
                 .append(this.publishDate)
@@ -263,7 +246,6 @@ public class Collection extends Content {
                 .append(this.shortUrl)
                 .append(this.status)
                 .append(this.taxonomy)
-                .append(this.title)
                 .appendSuper(super.hashCode())
                 .toHashCode();
     }
@@ -283,7 +265,7 @@ public class Collection extends Content {
                 .append(this.description, that.description)
                 .append(this.displayDate, that.displayDate)
                 .append(this.editorNote, that.editorNote)
-                .append(this.headline, that.headline)
+                .append(this.headlines, that.headlines)
                 .append(this.contentElements, that.contentElements)
                 .append(this.promoImages, that.promoImages)
                 .append(this.publishDate, that.publishDate)
@@ -291,7 +273,6 @@ public class Collection extends Content {
                 .append(this.shortUrl, that.shortUrl)
                 .append(this.status, that.status)
                 .append(this.taxonomy, that.taxonomy)
-                .append(this.title, that.title)
                 .appendSuper(super.equals(other))
                 .isEquals();
     }